refactor: moves api-related code out of hooks
