+++ src/hooks/use-playback.ts
@@ -3,31 +3,8 @@ import {
   useMutationState,
   useQueryClient,
 } from '@tanstack/react-query';
-import { getSession } from 'next-auth/react';
 
-const putPlayback = async (uris: string[]) => {
-  const session = await getSession();
-
-  if (!session) {
-    throw new Error('No Session');
-  }
-
-  const response = await fetch('https://api.spotify.com/v1/me/player/play', {
-    method: 'PUT',
-    headers: {
-      Authorization: `Bearer ${session.access_token}`,
-    },
-    body: JSON.stringify({
-      uris,
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error();
-  }
-
-  return;
-};
+import putPlayback from '@/lib/spotify/put-playback';
 
 const usePlayback = () => {
   const queryClient = useQueryClient();
