+++ src/hooks/use-queue.ts
@@ -1,29 +1,6 @@
-import { getSession } from 'next-auth/react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-const getAlbum = async (offset: number) => {
-  const session = await getSession();
-
-  if (!session) {
-    throw new Error('No Session');
-  }
-
-  const response = await fetch(
-    `https://api.spotify.com/v1/me/albums?offset=${offset}&limit=1&locale=*`,
-    {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${session.access_token}`,
-      },
-    },
-  );
-
-  if (!response.ok) {
-    throw new Error();
-  }
-
-  return (await response.json()) as Albums;
-};
+import getAlbums from '@/lib/spotify/get-albums';
 
 const useQueue = () => {
   const queryClient = useQueryClient();
@@ -33,11 +10,11 @@ const useQueue = () => {
     mutationFn: async () => {
       const totalSavedAlbums = await queryClient.ensureQueryData({
         queryKey: ['total-saved-albums'],
-        queryFn: async () => (await getAlbum(1)).total,
+        queryFn: async () => (await getAlbums(1, 1)).total,
       });
 
       const offset = Math.floor(Math.random() * totalSavedAlbums);
-      const response = await getAlbum(offset);
+      const response = await getAlbums(offset, 1);
 
       if (totalSavedAlbums !== response.total) {
         queryClient.setQueryData(['total-saved-albums'], () => response.total);
